chore: move search to sidebar
