+++ frontend/src/views/DashboardHeader.vue
@@ -2,28 +2,6 @@
   <div class="flex items-center justify-between h-10 pl-2 pr-4 space-x-2">
     <div class="flex-1 flex items-center">
       <BytebaseLogo class="block md:hidden" />
-
-      <div class="flex-1 hidden md:block">
-        <button
-          class="w-full max-w-md flex items-center justify-between rounded-md border border-control-border bg-white hover:bg-control-bg-hover pl-2 pr-1 py-0.5 outline-none"
-          @click="onClickSearchButton"
-        >
-          <span class="text-control-placeholder">
-            {{ $t("common.search") }}
-          </span>
-          <span class="flex items-center space-x-1">
-            <kbd
-              class="h-5 flex items-center justify-center bg-black bg-opacity-10 rounded text-sm px-1 text-control overflow-y-hidden"
-            >
-              <span v-if="isMac" class="text-xl px-0.5">⌘</span>
-              <span v-else class="tracking-tighter transform scale-x-90">
-                Ctrl
-              </span>
-              <span class="ml-1 mr-0.5">K</span>
-            </kbd>
-          </span>
-        </button>
-      </div>
     </div>
     <div>
       <div class="flex items-center space-x-3">
@@ -72,7 +50,6 @@
 
 <script lang="ts" setup>
 import { defineAction, useRegisterActions } from "@bytebase/vue-kbar";
-import { useKBarHandler } from "@bytebase/vue-kbar";
 import { Settings } from "lucide-vue-next";
 import { storeToRefs } from "pinia";
 import { computed, reactive, watchEffect } from "vue";
@@ -103,13 +80,6 @@ const { locale } = useLanguage();
 const state = reactive<LocalState>({
   showQRCodeModal: false,
 });
-
-const isMac = navigator.platform.match(/mac/i);
-const handler = useKBarHandler();
-
-const onClickSearchButton = () => {
-  handler.value.show();
-};
 
 const currentUser = useCurrentUser();
