+++ frontend/src/views/DashboardSidebar.vue
@@ -4,6 +4,24 @@
     <BytebaseLogo class="w-full px-4 shrink-0" />
 
     <div class="flex-1 overflow-y-auto px-2">
+      <button
+        class="mb-2 w-full flex items-center justify-between rounded-md border border-control-border bg-white hover:bg-control-bg-hover pl-2 pr-1 py-0.5 outline-none"
+        @click="onClickSearchButton"
+      >
+        <span class="text-control-placeholder">{{ $t("common.search") }}</span>
+        <span class="flex items-center space-x-1">
+          <kbd
+            class="h-5 flex items-center justify-center bg-black bg-opacity-10 rounded text-sm px-1 text-control overflow-y-hidden"
+          >
+            <span v-if="isMac" class="text-xl px-0.5">⌘</span>
+            <span v-else class="tracking-tighter transform scale-x-90"
+              >Ctrl</span
+            >
+            <span class="ml-1 mr-0.5">K</span>
+          </kbd>
+        </span>
+      </button>
+
       <router-link to="/" class="outline-item group flex items-center">
         <div
           class="outline-item group flex items-center px-2 py-1.5 capitalize"
@@ -95,6 +113,7 @@
 </template>
 
 <script lang="ts" setup>
+import { useKBarHandler } from "@bytebase/vue-kbar";
 import {
   HomeIcon,
   DatabaseIcon,
@@ -117,6 +136,8 @@ import {
 import { hasWorkspacePermissionV1 } from "../utils";
 
 const currentUserV1 = useCurrentUserV1();
+const isMac = navigator.platform.match(/mac/i);
+const handler = useKBarHandler();
 
 // Only show sync schema if the user has permission to alter schema of at least one project.
 const shouldShowSyncSchemaEntry = computed(() => {
@@ -128,6 +149,10 @@ const shouldShowSyncSchemaEntry = computed(() => {
     })
     .includes(true);
 });
+
+const onClickSearchButton = () => {
+  handler.value.show();
+};
 
 const shouldShowInstanceEntry = computed(() => {
   return hasWorkspacePermissionV1(
